requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
fastapi==0.94.1
h11==0.14.0
idna==3.4
jwt==1.3.1
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.6
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
